requirements.txt
pillow
numpy
pydicom
scikit-image
fastapi
uvicorn
python-multipart
requests
nibabel
nii2dcm
dicom2nifti
highdicom